+++ packages/ui-components/web/src/Components/SelectionList.tsx
@@ -179,16 +179,7 @@ class SelectionListOption extends Component<SelectionListOptionProps> {
   }
 
   render() {
-    const {
-      selected,
-      onChange,
-      children,
-      value,
-      // optionClassName,
-      dataHook,
-      tabIndex,
-      onKeyDown,
-    } = this.props;
+    const { selected, onChange, children, value, dataHook, tabIndex, onKeyDown } = this.props;
 
     return (
       <div
@@ -197,12 +188,6 @@ class SelectionListOption extends Component<SelectionListOptionProps> {
         aria-selected={selected}
         ref={el => (this.ref = el)}
         onKeyDown={e => onKeyDown?.(e)}
-        // className={
-        //   optionClassName ||
-        //   classnames(this.styles.selectionListOption, {
-        //     [this.styles.selectionListOption_selected]: selected,
-        //   })
-        // }
         data-hook={dataHook}
         onClick={() => onChange(value)}
       >